peripheral setup moved from seperate files into main.c
